+++ CHANGELOG.md
@@ -19,7 +19,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
-* `jj squash` now accepts `--editor` / `-E` to edit the squashed commit message.
+* `jj squash` and `jj split` now accept `--editor`, which ensures
+  an editor will be opened with the commit description even if one
+  was provided via `--message`/`-m`.
 
 ### Fixed bugs